+++ src/main/kotlin/controller/BlackjackController.kt
@@ -1,14 +1,15 @@
 package controller
 
-import entity.BettingMoney
-import entity.CardDistributeCondition
-import entity.Dealer
-import entity.Name
-import entity.Player
-import entity.Players
-import entity.PlayersGameResult
-import entity.UserBettingResult
-import entity.UserInformation
+import entity.result.BettingMoney
+import entity.card.CardDistributeCondition
+import entity.users.Dealer
+import entity.users.Name
+import entity.users.Player
+import entity.users.Players
+import entity.result.PlayersGameResult
+import entity.result.UserBettingResult
+import entity.users.UserInformation
+import entity.users.Users
 import model.BlackjackStage
 import model.RandomCardFactory
 import view.GameSetView
@@ -31,14 +32,14 @@ class BlackjackController {
         val players = readPlayersInformation()
         val dealer = Dealer()
         val cardFactory = RandomCardFactory()
-        val blackjackStage = BlackjackStage(dealer, players, cardFactory)
+        val blackjackStage = BlackjackStage(Users(players, dealer), cardFactory)
         blackjackStage.distributeAllUsers()
         gameView.printInitialUsersStatus(dealer, players)
         return blackjackStage
     }
 
     private fun distributeMoreCardPlayer(blackjackStage: BlackjackStage) {
-        blackjackStage.players.value.forEach { player ->
+        blackjackStage.users.players.value.forEach { player ->
             distributeMoreCardPlayerProcess(player, blackjackStage)
         }
     }
@@ -57,14 +58,14 @@ class BlackjackController {
     }
 
     private fun displayGameStatus(blackjackStage: BlackjackStage) =
-        resultView.printGameStatus(blackjackStage.dealer, blackjackStage.players)
+        resultView.printGameStatus(blackjackStage.users.dealer, blackjackStage.users.players)
 
     private fun getGameResult(blackjackStage: BlackjackStage) =
-        blackjackStage.players.determineAllPlayerGameResult(blackjackStage.dealer)
+        blackjackStage.users.players.determineAllPlayerGameResult(blackjackStage.users.dealer)
 
     private fun displayBettingResult(blackjackStage: BlackjackStage, playersGameResult: PlayersGameResult) {
         val userBettingResult = UserBettingResult()
-        val playersBettingResults = userBettingResult.getPlayersBettingResults(blackjackStage.players, blackjackStage.dealer, playersGameResult)
+        val playersBettingResults = userBettingResult.getPlayersBettingResults(blackjackStage.users.players, blackjackStage.users.dealer, playersGameResult)
         val dealerBettingResult = userBettingResult.getDealerBettingResult()
         resultView.printUserBettingResult(dealerBettingResult, playersBettingResults)
     }
