+++ src/main/kotlin/entity/card/Card.kt
@@ -1,3 +1,3 @@
-package entity
+package entity.card
 
 data class Card(val cardType: CardType, val cardNumber: CardNumber)
