+++ src/main/kotlin/entity/card/CardType.kt
@@ -1,4 +1,4 @@
-package entity
+package entity.card
 
 enum class CardType {
     SPADE, HEART, DIAMOND, CLUB
