+++ src/main/kotlin/entity/result/BettingMoney.kt
@@ -1,3 +1,3 @@
-package entity
+package entity.result
 
 data class BettingMoney(val value: Int)
