+++ src/main/kotlin/entity/result/DealerGameResult.kt
@@ -1,4 +1,4 @@
-package entity
+package entity.result
 
 class DealerGameResult(val value: Map<GameResultType, Int>) {
     init {
