+++ src/main/kotlin/entity/result/GameResultType.kt
@@ -1,4 +1,4 @@
-package entity
+package entity.result
 
 enum class GameResultType {
     WIN, DRAW, LOSE
