+++ src/main/kotlin/entity/users/Dealer.kt
@@ -1,4 +1,6 @@
-package entity
+package entity.users
+
+import entity.card.Cards
 
 class Dealer(cards: Cards = Cards(listOf())) : User(cards) {
     override fun isDistributable(): Boolean = cardsNumberSum() <= MAXIMUM_CARD_SUM_NUMBER
