+++ src/main/kotlin/entity/users/Player.kt
@@ -1,5 +1,7 @@
-package entity
+package entity.users
 
+import entity.result.GameResultType
+import entity.card.Cards
 import misc.GameRule
 
 class Player(val userInformation: UserInformation, cards: Cards = Cards(listOf())) : User(cards) {
