+++ src/main/kotlin/entity/users/UserInformation.kt
@@ -1,3 +1,5 @@
-package entity
+package entity.users
+
+import entity.result.BettingMoney
 
 class UserInformation(val name: Name, val bettingMoney: BettingMoney)
