+++ src/main/kotlin/entity/users/Users.kt
@@ -1,3 +1,3 @@
-package entity
+package entity.users
 
 data class Users(val players: Players, val dealer: Dealer)
