+++ src/main/kotlin/model/CardFactory.kt
@@ -1,6 +1,6 @@
 package model
 
-import entity.Cards
+import entity.card.Cards
 
 interface CardFactory {
     fun generate(count: Int): Cards
