+++ src/main/kotlin/view/ViewUtils.kt
@@ -1,9 +1,8 @@
 package view
 
-import entity.CardNumber
-import entity.CardType
-import entity.Cards
-import entity.GameResultType
+import entity.card.CardNumber
+import entity.card.CardType
+import entity.card.Cards
 
 class ViewUtils {
     companion object {
@@ -31,17 +30,5 @@ class ViewUtils {
                 it.cardNumber.isString() + it.cardType.isString()
             }
         }
-
-        fun GameResultType.isString(): String {
-            return when (this) {
-                GameResultType.WIN -> MESSAGE_WIN
-                GameResultType.LOSE -> MESSAGE_LOSE
-                GameResultType.DRAW -> MESSAGE_DRAW
-            }
-        }
-
-        private const val MESSAGE_WIN = "승"
-        private const val MESSAGE_LOSE = "패"
-        private const val MESSAGE_DRAW = "무"
     }
 }
