+++ src/test/kotlin/entity/CardNumberTest.kt
@@ -1,5 +1,6 @@
 package entity
 
+import entity.card.Cards
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
