+++ src/test/kotlin/entity/CardsTest.kt
@@ -1,5 +1,9 @@
 package entity
 
+import entity.card.Card
+import entity.card.CardNumber
+import entity.card.CardType
+import entity.card.Cards
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
