+++ src/test/kotlin/entity/NameTest.kt
@@ -1,5 +1,6 @@
 package entity
 
+import entity.users.Name
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
