+++ src/test/kotlin/entity/UserBettingResultTest.kt
@@ -1,5 +1,18 @@
 package entity
 
+import entity.card.Card
+import entity.card.CardNumber
+import entity.card.CardType
+import entity.card.Cards
+import entity.result.BettingMoney
+import entity.result.GameResultType
+import entity.result.PlayersGameResult
+import entity.result.UserBettingResult
+import entity.users.Dealer
+import entity.users.Name
+import entity.users.Player
+import entity.users.Players
+import entity.users.UserInformation
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
